move some actions out of menu
