+++ packages/plugins/plugin-space/src/capabilities/app-graph-builder.ts
@@ -14,6 +14,7 @@ import {
   SPACE_ID_LENGTH,
   SpaceState,
   type Space,
+  parseId,
 } from '@dxos/client/echo';
 import { isDeleted } from '@dxos/live-object';
 import { log } from '@dxos/log';
@@ -95,6 +96,45 @@ export default (context: PluginsContext) => {
             testId: 'spacePlugin.addSpace',
             disposition: 'item',
             position: 'fallback',
+          },
+        },
+        {
+          id: SpaceAction.Share._tag,
+          data: async () => {
+            const { dispatchPromise: dispatch } = context.requestCapability(Capabilities.IntentDispatcher);
+            const layout = context.requestCapability(Capabilities.Layout);
+            const client = context.requestCapability(ClientCapabilities.Client);
+            const { spaceId } = parseId(layout.workspace);
+            const space = (spaceId && client.spaces.get(spaceId)) ?? client.spaces.default;
+            await dispatch(createIntent(SpaceAction.Share, { space }));
+          },
+          properties: {
+            label: ['share space label', { ns: SPACE_PLUGIN }],
+            icon: 'ph--users--regular',
+            testId: 'spacePlugin.shareSpace',
+            keyBinding: {
+              macos: 'meta+.',
+              windows: 'alt+.',
+            },
+          },
+        },
+        {
+          id: SpaceAction.OpenSettings._tag,
+          data: async () => {
+            const { dispatchPromise: dispatch } = context.requestCapability(Capabilities.IntentDispatcher);
+            const layout = context.requestCapability(Capabilities.Layout);
+            const client = context.requestCapability(ClientCapabilities.Client);
+            const { spaceId } = parseId(layout.workspace);
+            const space = (spaceId && client.spaces.get(spaceId)) ?? client.spaces.default;
+            await dispatch(createIntent(SpaceAction.OpenSettings, { space }));
+          },
+          properties: {
+            label: ['open space settings label', { ns: SPACE_PLUGIN }],
+            icon: 'ph--gear--regular',
+            keyBinding: {
+              macos: 'meta+shift+,',
+              windows: 'ctrl+shift+,',
+            },
           },
         },
       ],
