+++ packages/plugins/plugin-space/src/components/MembersContainer.tsx
@@ -3,15 +3,15 @@
 //
 
 import { Check, X } from '@phosphor-icons/react';
-import React, { type Dispatch, type SetStateAction, useMemo, useState } from 'react';
+import React, { type Dispatch, type SetStateAction, useCallback, useMemo, useState } from 'react';
 import { QR } from 'react-qr-rounded';
 
 import { log } from '@dxos/log';
 import { useConfig } from '@dxos/react-client';
 import { fullyQualifiedId, useSpaceInvitations, type Space } from '@dxos/react-client/echo';
 import { type CancellableInvitationObservable, Invitation, InvitationEncoder } from '@dxos/react-client/invitations';
-import { Button, Clipboard, Icon, useId, useTranslation } from '@dxos/react-ui';
-import { ControlSection, ControlFrame, ControlFrameItem } from '@dxos/react-ui-form';
+import { Button, Clipboard, Icon, Input, useId, useTranslation } from '@dxos/react-ui';
+import { ControlSection, ControlFrame, ControlFrameItem, ControlItemInput } from '@dxos/react-ui-form';
 import { StackItem } from '@dxos/react-ui-stack';
 import { getSize, mx } from '@dxos/react-ui-theme';
 import {
@@ -28,6 +28,7 @@ import { hexToEmoji } from '@dxos/util';
 
 import { SPACE_PLUGIN } from '../meta';
 import { CollectionType } from '../types';
+import { COMPOSER_SPACE_LOCK } from '../util';
 
 // TODO(wittjosiah): Copied from Shell.
 const activeActionKey = 'dxos:react-shell/space-manager/active-action';
@@ -47,7 +48,7 @@ export const MembersContainer = ({
   space: Space;
   createInvitationUrl: (invitationCode: string) => string;
 }) => {
-  const { t } = useTranslation('os');
+  const { t } = useTranslation(SPACE_PLUGIN);
   const config = useConfig();
   const invitations = useSpaceInvitations(space.key);
   const visibleInvitations = invitations?.filter(
@@ -64,11 +65,16 @@ export const MembersContainer = ({
   // TODO(wittjosiah): Track which was the most recently viewed object.
   const target = space.properties[CollectionType.typename]?.target?.objects[0]?.target;
 
+  const locked = space.properties[COMPOSER_SPACE_LOCK];
+  const handleChangeLocked = useCallback(() => {
+    space.properties[COMPOSER_SPACE_LOCK] = !locked;
+  }, [locked, space]);
+
   const inviteActions = useMemo(
     (): Record<string, ActionMenuItem> => ({
       inviteOne: {
-        label: t('invite one label'),
-        description: t('invite one description'),
+        label: t('invite one label', { ns: 'os' }),
+        description: t('invite one description', { ns: 'os' }),
         icon: () => <Icon icon='ph--user-plus--regular' size={5} />,
         testId: 'membersContainer.inviteOne',
         onClick: () => {
@@ -86,8 +92,8 @@ export const MembersContainer = ({
         },
       },
       inviteMany: {
-        label: t('invite many label'),
-        description: t('invite many description'),
+        label: t('invite many label', { ns: 'os' }),
+        description: t('invite many description', { ns: 'os' }),
         icon: () => <Icon icon='ph--users-three--regular' size={5} />,
         testId: 'membersContainer.inviteMany',
         onClick: () => {
@@ -119,36 +125,46 @@ export const MembersContainer = ({
   return (
     <Clipboard.Provider>
       <StackItem.Content classNames='block overflow-y-auto'>
-        <ControlSection
-          title={t('members verbose label', { ns: SPACE_PLUGIN })}
-          description={t('members description', { ns: SPACE_PLUGIN })}
-        >
+        <ControlSection title={t('members verbose label')} description={t('members description')}>
           <ControlFrame>
-            <ControlFrameItem title={t('members label', { ns: SPACE_PLUGIN })}>
+            <ControlFrameItem title={t('members label')}>
               <SpaceMemberList spaceKey={space.key} includeSelf />
             </ControlFrameItem>
-            <ControlFrameItem title={t('invitations label', { ns: SPACE_PLUGIN })}>
-              {selectedInvitation && <InvitationSection {...selectedInvitation} onBack={handleBack} />}
-              {!selectedInvitation && (
-                <>
-                  <p className='text-description mbe-2'>{t('space invitation description', { ns: SPACE_PLUGIN })}</p>
-                  <InvitationList
-                    className='mb-2'
-                    send={handleSend}
-                    invitations={visibleInvitations ?? []}
-                    onClickRemove={(invitation) => invitation.cancel()}
-                    createInvitationUrl={createInvitationUrl}
-                  />
-                  <BifurcatedAction
-                    actions={inviteActions}
-                    activeAction={activeAction}
-                    onChangeActiveAction={setActiveAction as Dispatch<SetStateAction<string>>}
-                    data-testid='membersContainer.createInvitation'
-                  />
-                </>
-              )}
-            </ControlFrameItem>
+            {locked && (
+              <ControlFrameItem title={t('invitations label')}>
+                <p className='text-description mbe-2'>{t('locked space description')}</p>
+              </ControlFrameItem>
+            )}
+            {!locked && (
+              <ControlFrameItem title={t('invitations label')}>
+                {selectedInvitation && <InvitationSection {...selectedInvitation} onBack={handleBack} />}
+                {!selectedInvitation && (
+                  <>
+                    <p className='text-description mbe-2'>{t('space invitation description')}</p>
+                    <InvitationList
+                      className='mb-2'
+                      send={handleSend}
+                      invitations={visibleInvitations ?? []}
+                      onClickRemove={(invitation) => invitation.cancel()}
+                      createInvitationUrl={createInvitationUrl}
+                    />
+                    <BifurcatedAction
+                      actions={inviteActions}
+                      activeAction={activeAction}
+                      onChangeActiveAction={setActiveAction as Dispatch<SetStateAction<string>>}
+                      data-testid='membersContainer.createInvitation'
+                    />
+                  </>
+                )}
+              </ControlFrameItem>
+            )}
           </ControlFrame>
+          {/* TODO(wittjosiah): Make ControlItemInput & ControlFrame compatible. */}
+          <div className='justify-center gap-4 p-0 mbs-4 container-max-width grid grid-cols-1 md:grid-cols-[1fr_min-content]'>
+            <ControlItemInput title={t('space locked label')} description={t('space locked description')}>
+              <Input.Switch checked={locked} onCheckedChange={handleChangeLocked} classNames='justify-self-end' />
+            </ControlItemInput>
+          </div>
         </ControlSection>
       </StackItem.Content>
     </Clipboard.Provider>
