+++ packages/plugins/plugin-space/src/components/ObjectSettings/index.ts
