+++ packages/plugins/plugin-space/src/components/SpaceSettings/index.ts
@@ -2,5 +2,4 @@
 // Copyright 2024 DXOS.org
 //
 
-export * from './SpacePropertiesForm';
 export * from './SpaceSettingsContainer';
