+++ packages/plugins/plugin-space/src/translations.ts
@@ -34,7 +34,7 @@ export default [
         'export data label': 'Export data',
         'import data label': 'Import data',
         'migrate space label': 'Migrate space',
-        'share space': 'Share space',
+        'share space label': 'Share current space',
         'save space to disk label': 'Save to disk',
         'load space from disk label': 'Load from disk',
         'confirm restore title': 'Overwrite files in this space?',
@@ -43,7 +43,7 @@ export default [
         'object placeholder': 'Type a title here…',
         'personal space label': 'Personal Space',
         'spaces label': 'Spaces',
-        'show hidden spaces label': 'Show closed spaces',
+        'show hidden spaces label': 'Show archived spaces',
         'save files to directory label': 'Save files to disk',
         'select path label': 'Select path',
         'rename collection label': 'Rename collection',
@@ -67,7 +67,6 @@ export default [
           'The requested object has not been found yet. Ensure there are enough peers online in the space with an updated copy.',
         'create object in space label': 'Add to space',
         'create object in collection label': 'Add to collection',
-        'share space label': 'Share',
         'space members label': 'Space members',
         'active space members heading': 'Online ({{count}})',
         'inactive space members heading': 'Offline ({{count}})',
@@ -126,9 +125,12 @@ export default [
         'invitations label': 'Invitations',
         'space invitation description': 'Manage invitations to the space.',
         'qr code description': 'Share this QR code, or copy the URL and share it with the person you want to invite.',
+        'locked space description': 'This space is locked, unlock it to invite new members.',
+        'space locked label': 'Lock space',
+        'space locked description': 'Locking a space prevents new members from being invited to the space.',
 
         'settings panel label': 'Space Settings',
-        'open space settings label': 'Space Settings',
+        'open space settings label': 'Open current space settings',
         'space settings properties label': 'General',
         'space properties settings verbose label': 'Manage space properties',
         'space properties settings description': 'You can configure how this space is displayed in the app here.',
@@ -138,6 +140,10 @@ export default [
         'hue description': 'Color used to represent the space in the app.',
         'edge replication description':
           "Only change this if you know what you're doing. Disabling this will prevent the space from replicating through Composer's EDGE services, and relies solely on peer-to-peer sync.",
+        'archive space description':
+          'Archiving a space will remove it from the sidebar and stop replicating updates, but will not delete the data. You can unarchive it by enabling archived spaces in the app settings.',
+        'archive space label': 'Archive',
+        'unarchive space label': 'Unarchive',
       },
     },
   },
